+++ scripts/common.ts
@@ -3,6 +3,7 @@ import { ethers } from 'hardhat';
 import { Voucher, VoucherProxy } from '../../typechain-types';
 import { VoucherV2Mock } from '../../typechain-types/contracts/mocks';
 
+
 export async function getVoucher(voucherAddress: string): Promise<Voucher> {
     return await ethers.getContractAt('Voucher', voucherAddress);
 }