+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNEXT
+- Add role-based access control to VoucherHub. (#8)
 - Create voucher from VoucherHub with : type, expiration, authorize list. (#6)
 - Create vouchers with create2. (#5)
 - Create upgradeable voucher contract. (#4)